cas: Drop NfcNci Getting the following errors during build in A15 QPR2:
[ 85% 135603/158198] build out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/priv-app/NfcNci/NfcNci.apk
FAILED: out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/priv-app/NfcNci/NfcNci.apk
/bin/bash -c "(echo -e \"\\033[1mNfcNci: \\033[31merror:\\033[0m\\033[1m\" 'Module is requested to be installed but is not available for platform because it does not have \"//apex_available:platform\" or it depends on other modules that are not available for platform.'  \"\\033[0m\" >&2 ) && (exit 1 )"
NfcNci: error: Module is requested to be installed but is not available for platform because it does not have "//apex_available:platform" or it depends on other modules that are not available for platform.
[ 85% 135604/158198] build out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/priv-app/NfcNci/oat/arm64/NfcNci.odex
FAILED: out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/priv-app/NfcNci/oat/arm64/NfcNci.odex
/bin/bash -c "(echo -e \"\\033[1mNfcNci: \\033[31merror:\\033[0m\\033[1m\" 'Module is requested to be installed but is not available for platform because it does not have \"//apex_available:platform\" or it depends on other modules that are not available for platform.'  \"\\033[0m\" >&2 ) && (exit 1 )"
NfcNci: error: Module is requested to be installed but is not available for platform because it does not have "//apex_available:platform" or it depends on other modules that are not available for platform.
[ 85% 135605/158198] build out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/priv-app/NfcNci/oat/arm64/NfcNci.vdex
FAILED: out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/priv-app/NfcNci/oat/arm64/NfcNci.vdex
/bin/bash -c "(echo -e \"\\033[1mNfcNci: \\033[31merror:\\033[0m\\033[1m\" 'Module is requested to be installed but is not available for platform because it does not have \"//apex_available:platform\" or it depends on other modules that are not available for platform.'  \"\\033[0m\" >&2 ) && (exit 1 )"
NfcNci: error: Module is requested to be installed but is not available for platform because it does not have "//apex_available:platform" or it depends on other modules that are not available for platform.
[ 85% 135606/158198] build out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/etc/permissions/NfcNci.xml
FAILED: out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/etc/permissions/NfcNci.xml
/bin/bash -c "(echo -e \"\\033[1mNfcNci: \\033[31merror:\\033[0m\\033[1m\" 'Module is requested to be installed but is not available for platform because it does not have \"//apex_available:platform\" or it depends on other modules that are not available for platform.'  \"\\033[0m\" >&2 ) && (exit 1 )"
NfcNci: error: Module is requested to be installed but is not available for platform because it does not have "//apex_available:platform" or it depends on other modules that are not available for platform.
[ 85% 135607/158198] build out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/priv-app/NfcNci/lib/arm64/libnfc_nci_jni.so
FAILED: out/NOT_AVAILABLE_FOR_PLATFORM/target/product/cas/system/priv-app/NfcNci/lib/arm64/libnfc_nci_jni.so
/bin/bash -c "(echo -e \"\\033[1mNfcNci: \\033[31merror:\\033[0m\\033[1m\" 'Module is requested to be installed but is not available for platform because it does not have \"//apex_available:platform\" or it depends on other modules that are not available for platform.'  \"\\033[0m\" >&2 ) && (exit 1 )"
NfcNci: error: Module is requested to be installed but is not available for platform because it does not have "//apex_available:platform" or it depends on other modules that are not available for platform.
13:33:54 ninja failed with: exit status 1